+++ .NET/SalesWebMvc/SalesWebMvc/Models/Seller.cs
@@ -10,7 +10,7 @@ namespace SalesWebMvc.Models
         public string Name { get; set; }
         public string Email { get; set; }
         public DateTime BirthDate { get; set; }
-        public float BaseSalary { get; set; }
+        public double BaseSalary { get; set; }
         public Departament Departament { get; set; }
         public ICollection<SalesRecord> Sales { get; set; } = new List<SalesRecord>();
 
@@ -19,7 +19,7 @@ namespace SalesWebMvc.Models
 
         }
 
-        public Seller(int id, string name, string email, DateTime birthDate, float baseSalary, Departament departament)
+        public Seller(int id, string name, string email, DateTime birthDate, double baseSalary, Departament departament)
         {
             Id = id;
             Name = name;
@@ -39,7 +39,7 @@ namespace SalesWebMvc.Models
             Sales.Remove(sr);
         }
 
-        public float TotalSales(DateTime initial, DateTime final)
+        public double TotalSales(DateTime initial, DateTime final)
         {
             return Sales.Where(sr => sr.Date >= initial && sr.Date <= final).Sum(sr => sr.Amount);
         }
